[refactor] moving some `/settings` components to a more reasonable path
