+++ src/app/(app)/settings/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from "next";
 import { GetStartedChecklist } from "@components/system/get-started-checklist";
 import { getFeatureFlagWithPayload } from "src/core/utils/posthog-server-side";
 
-import { AutomationCodeReviewLayout } from "./code-review/_components/_layout";
+import { SettingsLayout } from "./_components/_layout";
 
 export const metadata: Metadata = {
     title: "Code Review Settings",
@@ -15,10 +15,9 @@ export default async function Layout({ children }: React.PropsWithChildren) {
     });
 
     return (
-        <AutomationCodeReviewLayout
-            pluginsPageFeatureFlag={pluginsPageFeatureFlag}>
+        <SettingsLayout pluginsPageFeatureFlag={pluginsPageFeatureFlag}>
             {children}
             <GetStartedChecklist />
-        </AutomationCodeReviewLayout>
+        </SettingsLayout>
     );
 }
